requirements.txt
aiogram
instagrapi
requests
apscheduler
aiosqlite
python-dotenv
aiohttp
aiofiles
